+++ slid-todo/src/app/(routes)/(todos)/todos/components/todo/index.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useState } from "react";
-import { useTodosInfinite } from "@/hooks/use-todos";
+import { useTodosInfinite } from "@/hooks/todo/use-todos";
 import TodoHeader from "./todo-header";
 import TodoFilter from "./todo-filter";
 import { TodoListContent } from "./todo-list-content";
